extracted directions to new class
