+++ lib/wrf_forecast/data.rb
@@ -1,0 +1,13 @@
+#!/usr/bin/ruby
+# @Author: Benjamin Held
+# @Date:   2020-08-06 19:32:21
+# @Last Modified by:   Benjamin Held
+# @Last Modified time: 2020-08-06 19:36:06
+
+module WrfForecast
+
+  require 'wrf_forecast/data/directions'
+  require 'wrf_forecast/data/forecast_repository'
+  require 'wrf_forecast/data/wind_direction_repository'
+
+end
