+++ lib/wrf_forecast/forecast/forecast_handler.rb
@@ -2,9 +2,9 @@
 # @Author: Benjamin Held
 # @Date:   2020-03-21 17:34:42
 # @Last Modified by:   Benjamin Held
-# @Last Modified time: 2020-04-21 17:46:33
+# @Last Modified time: 2020-08-06 19:34:08
 
-require 'wrf_forecast/data/forecast_repository'
+require 'wrf_forecast/data'
 require 'wrf_forecast/threshold'
 require 'wrf_forecast/forecast/forecast_text'
